split huge file to appropriate libs
