+++ ydb/core/tx/columnshard/operations/write.h
@@ -3,6 +3,7 @@
 #include <ydb/core/tx/data_events/write_data.h>
 #include <ydb/core/tx/data_events/events.h>
 #include <ydb/core/tx/columnshard/common/snapshot.h>
+#include <ydb/core/tx/columnshard/engines/scheme/versions/abstract_scheme.h>
 #include <ydb/core/tx/columnshard/engines/defs.h>
 #include <ydb/core/protos/tx_columnshard.pb.h>
