+++ ydb/core/tx/columnshard/operations/ya.make
@@ -3,7 +3,6 @@ LIBRARY()
 SRCS(
     write.cpp
     write_data.cpp
-    slice_builder.cpp
 )
 
 PEERDIR(
@@ -12,6 +11,8 @@ PEERDIR(
     ydb/services/metadata
     ydb/core/tx/columnshard/data_sharing/destination/events
     ydb/core/tx/columnshard/data_reader
+    ydb/core/tx/columnshard/operations/batch_builder
+    ydb/core/tx/columnshard/operations/slice_builder
 )
 
 END()
